// frontend/src/types/models.ts
export interface Employee {
  id: string
  clientId: string
  name: string
  pesel?: string
  contractType: 'UoP' | 'UZ'
  benefitAmount: number
}